+++ src/level.js
@@ -26,6 +26,7 @@ import { Sprite } from "kontra";
 import { playTune } from "./music.js";
 import { imageFromSvg, random } from "./utils.js";
 import { createEnemy } from "./enemy.js";
+import { createCloud, createLadder, createPlatform } from "./elements.js";
 import houseSvg from "./images/house.svg";
 
 const houseImage = imageFromSvg(houseSvg);
@@ -173,192 +174,14 @@ export class Level {
     context.fillRect(0, 0, this.width, this.height);
   }
 
-  createCloud(y, z, opacity) {
-    const level = this;
-
-    return Sprite({
-      x: Math.random() * level.width * (5 / 4) - level.width / 4,
-      y:
-        z === 0
-          ? y + (Math.random() * 200 - 102120)
-          : y + (Math.random() * 200 - 100),
-      color: "white",
-      opacity: z === 0 ? 0.95 * opacity : 0.7 * opacity,
-      dx: z === 0 ? 0.07 + Math.random() * 0.1 : 0.05 + Math.random() * 0.1,
-      radius: 20 + Math.random() * Math.random() * 70,
-
-      update: function() {
-        this.advance();
-        if (this.x - 300 > level.width) {
-          this.x = -600;
-        }
-      },
-
-      render: function() {
-        let cx = this.context;
-        cx.save();
-        cx.fillStyle = this.color;
-        cx.globalAlpha = this.opacity;
-        cx.beginPath();
-        let startX = this.x;
-        let startY = this.y;
-        cx.moveTo(startX, startY);
-        cx.bezierCurveTo(
-          startX - 40,
-          startY + 20,
-          startX - 40,
-          startY + 70,
-          startX + 60,
-          startY + 70
-        );
-        if (this.radius < 45) {
-          cx.bezierCurveTo(
-            startX + 80,
-            startY + 100,
-            startX + 150,
-            startY + 100,
-            startX + 170,
-            startY + 70
-          );
-        }
-        cx.bezierCurveTo(
-          startX + 250,
-          startY + 70,
-          startX + 250,
-          startY + 40,
-          startX + 220,
-          startY + 20
-        );
-        cx.bezierCurveTo(
-          startX + 260,
-          startY - 40,
-          startX + 200,
-          startY - 50,
-          startX + 170,
-          startY - 40
-        );
-        cx.bezierCurveTo(
-          startX + 150,
-          startY - 75,
-          startX + 80,
-          startY - 60,
-          startX + 80,
-          startY - 40
-        );
-        cx.bezierCurveTo(
-          startX + 30,
-          startY - 75,
-          startX - 20,
-          startY - 60,
-          startX,
-          startY
-        );
-
-        cx.closePath();
-        cx.fill();
-        cx.restore();
-      }
-    });
-  }
-
-  createLadder() {
-    return Sprite({
-      color: "rgb(100,60,60)",
-      color2: "rgb(80,20,20)",
-      width: 30,
-      height: 100,
-      stepGap: 15,
-
-      render: function() {
-        const stepCount = this.height / this.stepGap + 1;
-        let cx = this.context;
-        cx.save();
-
-        for (let i = 0; i < stepCount; i++) {
-          cx.fillStyle = this.color2;
-          cx.fillRect(
-            this.x + 8,
-            this.y + i * this.stepGap,
-            this.width - 16,
-            this.stepGap / 2
-          );
-          cx.fillStyle = this.color;
-          cx.fillRect(
-            this.x,
-            this.y + i * this.stepGap + this.stepGap / 2,
-            this.width,
-            this.stepGap / 2
-          );
-        }
-
-        cx.restore();
-      }
-    });
-  }
-
-  createPlatform(isBackground) {
-    return Sprite({
-      color: "darkgray",
-      color2: "rgb(80,80,80)",
-      color3: "rgb(55,55,75)",
-      width: 200,
-      height: !isBackground ? 20 : 400,
-      opacity: 0.1,
-
-      render: function() {
-        const stepGap = 20;
-        const stepCount = this.height / stepGap;
-        let cx = this.context;
-        cx.save();
-        if (!isBackground) {
-          for (let i = 0; i < stepCount; i++) {
-            cx.fillStyle = this.color2;
-            cx.fillRect(this.x + 2, this.y + i * stepGap, this.width - 4, 10);
-            cx.fillStyle = this.color;
-            cx.fillRect(this.x, this.y + i * stepGap + 15, this.width, 10);
-          }
-        } else {
-          cx.globalAlpha = this.opacity;
-          cx.fillStyle = this.color3;
-          cx.fillRect(this.x, this.y, this.width, this.height);
-          cx.strokeStyle = this.color3;
-          cx.lineWidth = 1;
-          cx.strokeRect(this.x, this.y, this.width, this.height);
-        }
-
-        cx.restore();
-      }
-    });
-  }
-
-  createRoof() {
-    const level = this;
-
-    return Sprite({
-      color: "darkgray",
-      width: level.width,
-      height: 30,
-      x: 0,
-      y: level.height - 30,
-
-      render: function() {
-        let cx = this.context;
-        cx.save();
-        cx.fillStyle = this.color;
-        cx.fillRect(this.x, this.y, this.width, this.height);
-        cx.restore();
-      }
-    });
-  }
-
   createCloudLayer(y, opacity) {
     for (let i = 0; i < 20; i++) {
-      let cloud0 = this.createCloud(y, 0, opacity);
+      let cloud0 = createCloud(this, y, 0, opacity);
       this.clouds0.push(cloud0);
     }
 
     for (let i = 0; i < 20; i++) {
-      let cloud1 = this.createCloud(y, 1, opacity);
+      let cloud1 = createCloud(this, y, 1, opacity);
       this.clouds1.push(cloud1);
     }
   }
@@ -393,7 +216,7 @@ export class Level {
   }
 
   addLadder(platform, height) {
-    let ladder = this.createLadder();
+    let ladder = createLadder();
     ladder.height = height;
     ladder.x = platform.x + 20;
     ladder.y = platform.y;
@@ -414,7 +237,7 @@ export class Level {
       // No two consecutive platforms with holes in them.
       if (isHoleOnPreviousLayer || random() < 0.8) {
         // One solid platform
-        let platform = this.createPlatform(false);
+        let platform = createPlatform(false);
         platform.width = floorWidth;
         platform.x = floorLeft;
         platform.y = floorTop;
@@ -434,14 +257,14 @@ export class Level {
         // Two platforms and a hole between them.
         const lessWidth = floorWidth / 3;
 
-        let p1 = this.createPlatform(false);
+        let p1 = createPlatform(false);
         p1.width = lessWidth;
         p1.x = floorLeft;
         p1.y = floorTop;
         this.platforms.push(p1);
         this.addLadder(p1, floorHeight);
 
-        let p2 = this.createPlatform(false);
+        let p2 = createPlatform(false);
         p2.width = lessWidth;
         p2.x = floorRight - lessWidth;
         p2.y = floorTop;
@@ -450,7 +273,7 @@ export class Level {
         isHoleOnPreviousLayer = true;
       }
 
-      let platformBg = this.createPlatform(true);
+      let platformBg = createPlatform(true);
       platformBg.width = floorWidth;
       platformBg.x = floorLeft;
       platformBg.y = floorTop;
